server/requirements.txt
fastapi
uvicorn
anthropic
python-dotenv
pydantic
httpx
python-multipart
pillow
PyPDF2
python-docx
pandas
openpyxl
chardet